Introduce IOStub and resolve remaining specs
